feat(rpc): Add eth_chainId and eth_blockNumber
