+++ rpc/functions.go
@@ -7,28 +7,8 @@ import (
 	tmrpccore "github.com/tendermint/tendermint/rpc/core"
 	tmrpccoretypes "github.com/tendermint/tendermint/rpc/core/types"
 	tmrpctypes "github.com/tendermint/tendermint/rpc/jsonrpc/types"
-	"regexp"
 	"strings"
 )
-
-var hexReg = regexp.MustCompile(`(?i)[a-f0-9]{40,}`)
-
-func parseHeight(heightPtr *int64) int64 {
-	if heightPtr == nil {
-		return 0
-	}
-	return *heightPtr
-}
-
-func parsePath(ctx *tmrpctypes.Context) string {
-	if ctx.JSONReq != nil {
-		return ctx.JSONReq.Method
-	}
-	if ctx.HTTPReq != nil {
-		return strings.TrimSuffix(strings.TrimPrefix(ctx.HTTPReq.URL.Path, "/"), "/")
-	}
-	return ""
-}
 
 func QueryAccount(ctx *tmrpctypes.Context, addr abytes.HexBytes, heightPtr *int64) (*QueryResult, error) {
 	height := parseHeight(heightPtr)
