+++ rpc/utils.go
@@ -1,0 +1,26 @@
+package rpc
+
+import (
+	tmrpctypes "github.com/tendermint/tendermint/rpc/jsonrpc/types"
+	"regexp"
+	"strings"
+)
+
+var hexReg = regexp.MustCompile(`(?i)[a-f0-9]{40,}`)
+
+func parseHeight(heightPtr *int64) int64 {
+	if heightPtr == nil {
+		return 0
+	}
+	return *heightPtr
+}
+
+func parsePath(ctx *tmrpctypes.Context) string {
+	if ctx.JSONReq != nil {
+		return ctx.JSONReq.Method
+	}
+	if ctx.HTTPReq != nil {
+		return strings.TrimSuffix(strings.TrimPrefix(ctx.HTTPReq.URL.Path, "/"), "/")
+	}
+	return ""
+}
